Convert Badge to Typescript
