--- packages/mui-base/src/BadgeUnstyled/index.js
@@ -1,6 +1,0 @@
-export { default } from './BadgeUnstyled';
-export { default as useBadge } from './useBadge';
-export {
-  default as badgeUnstyledClasses,
-  getBadgeUnstyledUtilityClass,
-} from './badgeUnstyledClasses';
