+++ packages/mui-base/src/BadgeUnstyled/index.ts
@@ -4,8 +4,7 @@ export * from './BadgeUnstyled';
 export { default as useBadge } from './useBadge';
 export * from './useBadge';
 
-export { default as BadgeUnstyledProps } from './BadgeUnstyledProps';
-export * from './BadgeUnstyledProps';
+export * from './BadgeUnstyled.types';
 
 export { default as badgeUnstyledClasses } from './badgeUnstyledClasses';
 export * from './badgeUnstyledClasses';
